requirements.txt
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2017.3
django-recaptcha==1.3.1
django-form-utils==1.0.3
psycopg2==2.7.3.2
django-loginas==0.3.3
djangorestframework-jwt==1.11.0
 django-extensions-1.9.9
